codex: revert unrelated CI fixes on PR #18632
