+++ codex-rs/tui/src/chatwidget.rs
@@ -2191,11 +2191,10 @@ impl ChatWidget {
 
     pub(crate) fn handle_thread_session_quiet(&mut self, session: ThreadSessionState) {
         self.instruction_source_paths = session.instruction_source_paths.clone();
-        let fork_parent_title = session.fork_parent_title.clone();
         self.on_session_configured_with_display_and_fork_parent_title(
             thread_session_state_to_legacy_event(session),
             SessionConfiguredDisplay::Quiet,
-            fork_parent_title,
+            /*fork_parent_title*/ None,
         );
     }
